icaas/local_requirements.txt
tensorflow==1.14.0
python-decouple==3.1
requests==2.22.0
bert-serving-server==1.9.8
bert-serving-client==1.9.8